cleanup

+++ src/renderer/editor/particle_editor.cpp
@@ -1170,15 +1170,15 @@ struct ParticleEditorImpl : ParticleEditor {
 		m_toggle_ui.func.bind<&ParticleEditorImpl::toggleOpen>(this);
 		m_toggle_ui.is_selected.bind<&ParticleEditorImpl::isOpen>(this);
 
-		m_undo_action.init(ICON_FA_UNDO "Undo", "Particle editor undo", "particle_editor_undo", ICON_FA_UNDO, os::Keycode::Z, (u8)Action::Modifiers::CTRL, true);
+		m_undo_action.init(ICON_FA_UNDO "Undo", "Particle editor undo", "particle_editor_undo", ICON_FA_UNDO, os::Keycode::Z, Action::Modifiers::CTRL, true);
 		m_undo_action.func.bind<&ParticleEditorImpl::undo>(this);
 		m_undo_action.plugin = this;
 
-		m_redo_action.init(ICON_FA_REDO "Redo", "Particle editor redo", "particle_editor_redo", ICON_FA_REDO, os::Keycode::Z, (u8)Action::Modifiers::CTRL |  (u8)Action::Modifiers::SHIFT, true);
+		m_redo_action.init(ICON_FA_REDO "Redo", "Particle editor redo", "particle_editor_redo", ICON_FA_REDO, os::Keycode::Z, Action::Modifiers::CTRL | Action::Modifiers::SHIFT, true);
 		m_redo_action.func.bind<&ParticleEditorImpl::redo>(this);
 		m_redo_action.plugin = this;
 
-		m_apply_action.init("Apply", "Particle editor apply", "particle_editor_apply", "", os::Keycode::E, (u8)Action::Modifiers::CTRL, true);
+		m_apply_action.init("Apply", "Particle editor apply", "particle_editor_apply", "", os::Keycode::E, Action::Modifiers::CTRL, true);
 		m_apply_action.func.bind<&ParticleEditorImpl::apply>(this);
 		m_apply_action.plugin = this;
